requirements.txt
flask>=3.0
PyYAML>=5.0
